Move `HistoryWrapper` to the `wrapper` module
